building path recursively
