+++ internal/redact/redact.go
@@ -12,7 +12,7 @@ import (
 // Redact removes the sensitive keys provided in remote config cache
 // NOTE: Redact modifies events and appends redacted info to the event object
 // NOTE: Redact is expecting that the endpoint Id for the event has been successfully populated
-// during event creation
+// during event
 func Redact(events []*event.Event, rc *remoteconfig.RemoteConfig) []error {
 	var errs []error
 	for _, e := range events {
@@ -32,7 +32,6 @@ func Redact(events []*event.Event, rc *remoteconfig.RemoteConfig) []error {
 				errs = append(errs, err)
 				continue
 			}
-			// TODO: pass down empty string and build out path as i traverse
 			meta, err := redactEvent(sensitiveKey.KeyPath, formattedParts, e)
 			if err != nil {
 				errs = append(errs, err)
@@ -45,10 +44,15 @@ func Redact(events []*event.Event, rc *remoteconfig.RemoteConfig) []error {
 }
 
 func redactEvent(fullpath string, path []string, v any) ([]event.RedactedKeyMeta, error) {
-	return redactEventHelper(fullpath, path, reflect.ValueOf(v).Elem())
+	return redactEventHelper(fullpath, path, reflect.ValueOf(v).Elem(), "")
 }
 
-func redactEventHelper(fullpath string, path []string, v reflect.Value) ([]event.RedactedKeyMeta, error) {
+// redactEventHelper recursively redacts fields based off a provided path
+// fullpath: represents the original full path - used for logging
+// path: is the split stringed representation of the fullpath. contains the remaining path elements to be traversed
+// v: is the reflected value at the current path
+// createdPath: is the recursively built path. Difference between fullpath and createdpath is that createdPath will contain array indexes
+func redactEventHelper(fullpath string, path []string, v reflect.Value, createdPath string) ([]event.RedactedKeyMeta, error) {
 	if !v.IsValid() {
 		return nil, fmt.Errorf("unable to find key at sensitive key provided path: %s", fullpath)
 	}
@@ -60,7 +64,7 @@ func redactEventHelper(fullpath string, path []string, v reflect.Value) ([]event
 		v.Set(reflect.Zero(v.Type()))
 		return []event.RedactedKeyMeta{
 			{
-				KeyPath: fullpath,
+				KeyPath: reformatSensitiveKeyPath(createdPath),
 				Length:  size,
 				Type:    v.Type().Kind().String(),
 			},
@@ -68,10 +72,10 @@ func redactEventHelper(fullpath string, path []string, v reflect.Value) ([]event
 	} else {
 		switch v.Type().Kind() {
 		case reflect.Ptr, reflect.Interface:
-			return redactEventHelper(fullpath, path, v.Elem())
+			return redactEventHelper(fullpath, path, v.Elem(), createdPath)
 
 		case reflect.Struct:
-			return redactEventHelper(fullpath, path[1:], v.FieldByName(path[0]))
+			return redactEventHelper(fullpath, path[1:], v.FieldByName(path[0]), createdPath+"."+path[0])
 
 		case reflect.Map:
 			// You can't mutate elements of a map, but as a special case if we just
@@ -92,32 +96,30 @@ func redactEventHelper(fullpath string, path []string, v reflect.Value) ([]event
 				v.SetMapIndex(idx, reflect.Zero(v.Type().Elem()))
 				return []event.RedactedKeyMeta{
 					{
-						KeyPath: fullpath,
+						KeyPath: reformatSensitiveKeyPath(formatFieldPathPart(createdPath, path[0])),
 						Length:  size,
 						Type:    objKind,
 					},
 				}, nil
 			} else {
-				return redactEventHelper(fullpath, path[1:], mapVal)
+				return redactEventHelper(fullpath, path[1:], mapVal, formatFieldPathPart(createdPath, path[0]))
 			}
 
 		case reflect.Array, reflect.Slice:
 			idx := parseArrayIndex(path[0])
-			if idx > -1 {
-				return redactEventHelper(fullpath, path[1:], v.Index(idx))
-			} else if idx == -1 {
-				results := []event.RedactedKeyMeta{}
-				for i := 0; i < v.Len(); i++ {
-					result, err := redactEventHelper(fullpath, path[1:], v.Index(i))
-					if err != nil {
-						return results, err
-					}
-					results = append(results, result...)
-				}
-				return results, nil
-			} else {
+			if idx == -1 {
 				return nil, fmt.Errorf("invalid index value provided at location")
 			}
+
+			results := []event.RedactedKeyMeta{}
+			for i := 0; i < v.Len(); i++ {
+				result, err := redactEventHelper(fullpath, path[1:], v.Index(i), formatArrayPathPart(createdPath, i))
+				if err != nil {
+					return results, err
+				}
+				results = append(results, result...)
+			}
+			return results, nil
 
 		default:
 			return nil, fmt.Errorf("redact.Redact: unsupported type %v", v.Type().String())
