+++ internal/redact/size.go
@@ -1,0 +1,28 @@
+package redact
+
+import "reflect"
+
+// Note: this is a naive way of generating the size of a reflected object
+func getSize(v reflect.Value) int {
+	size := 0
+	switch v.Kind() {
+	case reflect.Interface, reflect.Pointer:
+		size += getSize(v.Elem())
+	case reflect.Array, reflect.Slice:
+		for i := 0; i < v.Len(); i++ {
+			size += getSize(v.Index(i))
+		}
+	case reflect.Map:
+		keys := v.MapKeys()
+		for i := range keys {
+			size += getSize(keys[i]) + getSize(v.MapIndex(keys[i]))
+		}
+	case reflect.String:
+		size += v.Len()
+	case reflect.Struct:
+		for i := 0; i < v.NumField(); i++ {
+			size += getSize(v.Field(i))
+		}
+	}
+	return size
+}
